Created 2 files, utils for methods in frontend, and plots for plotting methods, these are now imported into the dashboard-files
