+++ pa_agent/src/pa_agent/data/dagbok.csv
@@ -67,5 +67,5 @@ date,weekday,activity,feelings,mood
 2026-01-22,Torsdag,Tänkte på ord och försökte förstå,Tankfull,positivt
 2026-01-22,Torsdag,Något är på gång och jag hoppas det fungerar,Glad,positivt
 2026-01-26,Måndag,Kört barnen till skolan,Avslappnad,positivt
-2026-01-26,Måndag,Lek med ungarna,Skönt,Positivt
-2026-01-27,Tisdag,pluggat,"jättekul, pepp",Positive
+2026-01-26,Måndag,Lek med ungarna,Skönt,positivt
+2026-01-27,Tisdag,pluggat,"jättekul, pepp",positivt
